deleted copyright definition and entity transactions provision
In 2006 lawyers for Sun Microsystems suggested adding an explicit
definition of "copyright" that would encompass not just conventional
copyright law but also laws like the US Semiconductor Chip Protection
Act (which is something of a hybrid of copyright and patent law). The
application of the GNU GPLv3 to semiconductor integrated circuit
layouts proved to be an unimportant use case. No other
widely-encountered free software license attempts explicitly to apply
its provisions beyond copyright to such "copyright-like laws". In the
interest of brevity I have therefore deleted this definition.

I have deleted the middle paragraph of section 10, the "entity
transactions" provision. This originated in a provision drafted by
Eben Moglen as an alternative to a suggestion by Mark Radcliffe that
GPLv3 have an explicit assignability provision to address a concern
believed by some lawyers to arise in corporate acquisitions involving
GPL-licensed software assets. I am currently inclined to regard Mark's
concern as not being a real-world problem, and perhaps unnecessary
given the automatic licensing feature of the GPL but at any rate I
consider a simple conventional assignability provision preferable to
the existing entity transactions provision, which seems more complex
than necessary. It is not clear how the "right to possession of the
Corresponding Source" is to be enforced (it might be simpler to say
that an entity transaction is a form of conveying, and in fact I seem
to recall some interim GPLv3 draft language taking such an
approach). In any case, this is yet another provision that finds no
counterpart in other major free software licenses. In the interest of
greater brevity it should be deleted, but the issue of assignment in
the M&A context should perhaps be revisited at a later time.