restored original name of present-day sec. 12
What is now section 12 of GPLv3 (the successor to GPLv2 section 7)
bore the heading "Liberty or Death for the Program" in the initial
public draft of GPLv3. The heading was, I believe, devised by
RMS. This wonderful title was, I think, thought to be too
radical-sounding or something (perhaps a vendor or commercial user
raised such a concern early on during the GPLv3 drafting process), and
so the drafters changed it to "No Surrender of Others' Freedom". The
original heading is clearly superior. I have myself taken to calling
this provision the "liberty-or-death clause" and such terminology has
to some degree caught on with others.

In fairness, I recall that RMS also made the point that under this
provision the Program does not actually "die"; rather, particular
copies of it may become nondistributable. Thus the heading was never
truly accurate.  But the principal reason for this change was
lamentable primness.

I have therefore restored the original heading.