deleted definition of License; removed regeneration clause from definition of Corresponding Source
This commit makes two small changes. First, the explicit definition of
"License" is deleted as *probably* unnecessary. In GPLv2 the general
understanding has been that "License" obviously means GPLv2. Now I
recall that adding an explicit definition of "License" in GPLv3 (which
I believe was my idea) did serve some purpose other than
hyperlegalism. I believe it may have been thought desirable to clarify
that in *any* version of the GNU GPL "this License" cannot mean "some
or all versions of the GNU GPL" except in the special case of the
"revised versions" provision. (As a concrete example, consider the
meaning of GPLv2 section 2 for a "GPLv2 only" program: if "this
License" could mean "any version of the GPL", it would frustrate the
expectations of the licensors of the GPLv2-only program that modified
versions be licensed only under GPLv2 and not GPLv3 or some other
version.) Perhaps this should be revisited, but there is essentially
no basis in real-world experience that justifies the specific fear
that would motivate this clarification.

Admittedly, the GNU GPL family is unusual as a copyleft license that
does not have a built-in "or later" clause, and perhaps this feature
or non-feature (clearly important to many GPL licensors of existing
projects, such as the Linux kernel) justifies the clarification about
what "this License" means. Nevertheless in five years of experience of
GPLv2 no one has attempted to argue that because "License" is
undefined in GPLv2 one can license out a modified version of a
GPLv2-only program under GPLv3. (The counterargument to that may be
that GPLv3 retroactively provides that clarification for GPLv2.)

The second change in this commit is deletion of the following
paragraph from the definition of Corresponding Source:

  The Corresponding Source need not include anything that users can
  regenerate automatically from other parts of the Corresponding
  Source.

This must have been added for a reason, but I do not believe there is
any record of it. I do not believe I was involved in discussing or
drafting it (it seems likely that it was triggered by a stet
comment). I'm sure there is some technical circumstance in which this
sentence will provide some marginally desirable clarification of the
limits of the Corresponding Source requirement but it is not obvious
to me what such a circumstance would be. The definition of
Corresponding Source is already complex enough. GPLv2 has gotten by
without this clarification. In my post-2007 experience of GPLv3 I have
never been aware of any issue triggering this clause. In the interest
of brevity and simplification I have decided to delete it. If someone
would like to explain what it is aimed at and why explicit inclusion
of the clause is more helpful than hurtful (in making the
Corresponding Source definition more complex) I'm all ears.