deleted separable-portion clause from section 6
This commit deletes from section 6 the following paragraph:

  A separable portion of the object code, whose source code is
  excluded from the Corresponding Source as a System Library, need not
  be included in conveying the object code work.

I do not recall the reason for inclusion of this clause, though it may
have been a response to a stet comment. It is a bit strange in that
(the first half of) section 6 is concerned with the requirement to
convey Corresponding Source triggered by the action of conveying an
object code version of a work, yet this clause reads like a
clarification of a requirement to convey the object code. Of course
there is no requirement to convey object code.

What I suppose this clause must be saying is that where the
distributor ordinarily would not be conveying the System Library with
its object code work, the fact that one is required to convey
Corresponding Source (modulo System Libraries) does not somehow then
imply a requirement to convey the object code System Library. I can't
see how anyone would reasonbly think this could possibly be a
requirement in a provision that says nothing otherwise about a
requirement to provide object code.