simplified title of sec. 3
This commit simplifies the heading of section 3 to reduce it from
political chargedness to descriptiveness (without affecting the
content of the provision).

Historically, section 3 of GPLv3 was called "Digital Restrictions
Management", the FSF's reexpansion of DRM. Over time the provision,
once quite broad in scope, came to focus entirely on
anti-circumvention law.