greatly simplified acceptance provision
Section 9 of GPLv3 originated as a provision titled "Not a
Contract". The mere title resulted in some objections by some European
lawyers and at least bemusement on the part of some American
lawyers. Curiously, though, the GPLv3 provision and its GPLv2
predecessor deal with acceptance, which is a contract law concept.

I believe the ancient "license vs. contract" debates concerning the
GPL were characterized by great confusion on all sides. There are some
interesting legal issues related to this which may be worth further
exploration. At this time, however, I accept the idea that the GPL
should have some sort of acceptance clause (as indeed many proprietary
software licenses and some free software licenses do), if only for
reasons of tradition, and that the FSF may intend for it to serve some
legal purpose that has nothing to do with acceptance in the contract
sense (or perhaps really does wish it to have some contract law
effect, traditional insistence that the GPL is "not a contract"
notwithstanding). What I have done here is simplified the title to
"Acceptance" and reduced the content of the provision to simply state
that modification or propagation are deemed to signify acceptance by
the licensee. I have kept the policy but trimmed the language to the
bare essentials.