deleted MS coupon provision
This commit deletes the paragraph in section 11 which was intended, as
a hack on a feature of the Microsoft/Novell deal of 2006, to cause
Microsoft patent licenses/covenants granted to Novell customers to be
"automatically extended" to all downstream recipients of the software
associated with the SLES certificates distributed by Microsoft
pursuant to the deal.

Historical evidence shows that this provision was taken somewhat
seriously, and its cleverness and creativity are to be
appreciated. However, it must be admitted that it has served no valid
purpose in the past five years. To my knowledge no one has ever
attempted to invoke the provision (or had need to invoke it) to argue
for the existence of a patent license or covenant, and it would be
exceedingly strange for anyone to have done so. The provision is
worded generally, but it is tied to one deal between two specific
companies and is intended to punish one particular company. Whatever
its political value was in early 2007, the provision today is either a
no-op or serves to intensify anti-GPLv3 FUD. It needs to go.

One can view this provision as a sort of odd exception to the
historical narrowing of the general patent license grant now contained
in paragraphs 1-3 of section 11. Early public drafts of GPLv3 had
featured a "pure distribution" approach to patent licensing. Whether
that narrowing of the patent license grant was good policy or not is
an open question, but it is worth noting that this anti-Microsoft
provision would have been pointless had a pure-distribution patent
licensing policy been retained.