modified final sentence of sec. 7
This commit simplifies and clarifies the final sentence/paragraph of
section 7. The concern here was uncertainty over the extent to which
the section 7 "allowed additional requirements" went beyond the
limited setting of compatibility with discrete free software
licenses. For example, section 7 notes that it is not a violation of
section 10 to supplement GPLv3 with a differently-worded warranty
disclaimer. The final sentence makes clear that this not only ratifies
the compatibility of differently-worded warranty disclaimers in
licenses traditionally treated as GPL-compatible, but also authorizes
an informal supplementation of the GPLv3 warranty disclaimer.

I changed "or stated as exceptions" to "or stated as exceptions to or
qualifications of", because the term "exception" in GPL contexts has
generally referred only to additional permissions, and application of
the term "exception" to an additional restriction seems confusing, at
least in light of this terminology tradition. (Cf. the original
license of Liberation Fonts.)