deleted section 17
This commit deletes a late change to GPLv3, the final section, which
had been drafted in response to arguments made by Axel Metzger
concerning the effects of the US-style warranty and liability
disclaimers under German law. While I like this provision, I must
wonder why it is necessary given that GPLv2 and other widely-used free
software licenses have gotten along fine without anything like it
(similar issues must arise with proprietary software licenses). It
should also be noted that Metzger's concerns may have resonance in US
law. Perhaps the better approach is to attempt improvement of the
disclaimers themselves, and/or to reconsider the idea, rejected early
on in the drafting of GPLv3, of a general severability clause.