Initial draft
I deleted the Preamble of the GNU GPLv3. This was necessary because
the FSF has stated this as a requirement for permission to create
derivatives of the GNU GPL. I could write an alternative Preamble, but
for the moment I am inclined to think that GPL.next does not need
one.

The GNU GPL Preamble (in its various versions) is an inspiring and
important political statement and also functions a bit similarly to
the so-called "human-readable code" counterparts to Creative Commons
licenses in providing a basic summary description of the license (or
at any rate that is one intention behind the GNU GPL
Preamble). However, even if one thinks a license summary is valuable,
it may be better as a separate document (the Creative Commons
approach).

As for the political aspect of the Preamble, there is something
problematic about this in that it expresses the viewpoint of the FSF
and its sympathizers, but most likely does not quite speak for a
substantial number (perhaps the majority) of GPL licensors (and
licensees). Developers who disagree with the FSF on certain policy
issues (but who favor or potentially favor, strong copyleft) may find
a Preamble-less GPL less alienating.

The FSF also requires GNU GPL derivatives to reword the "How to Apply"
appendix and remove references to GNU, if the appendix is preserved at
all. I decided simply to delete it, particularly in view of the length
of the license document. Instructions on how to apply the license may
be useful but they need not be a part of the license document. Certain
interim public drafts of the GNU GPLv3 removed the appendix for this
reason, though it was later restored.

Section 13 of the GNU GPLv3 and a corresponding paragraph in section
13 of the GNU AGPLv3 provide for cross-compatibility of these two
licenses. I have replaced this section with a quickly-drafted sentence
that says that GPL.next can be relicensed under any version of the GNU
GPL or GNU AGPL (thus making it GNU GPL-compatible, which addresses
one of the FSF's stated concerns regarding GPL derivatives).

I substituted "License Steward" for the references to the FSF in
section 14, and (awkwardly, I later decided) declared myself to be the
License Steward "for the time being".