Restores warranty disclaimer and limitation of liability provisions to lowercase.
The initial public draft of GPLv3 changed the all-caps disclaimers of
GPLv3 to lowercase (more precisely mixed-case) out of concern that
there was no good reason for "shouting". Subsequently, some US lawyers
pointed out the requirement under the Uniform Commercial Code (where
it is applicable) for at least portions of such disclaimers to be
"conspicuous". As noted in the GPLv3 Discussion Draft 3 Rationale:

  There is authority under United States law suggesting that effective
  warranty disclaimers must be “conspicuous,” and that conspicuousness
  can be established by capitalization and is absent when a disclaimer
  has the same typeface as the terms surrounding it (see Stevenson
  v. TRW, Inc., 987 F.2d 288, 296 (5th Cir. 1993)). We have reason to
  doubt that such authority would apply to copyright licenses like the
  GPL. Nevertheless, pending further research, we have cautiously
  decided to restore the capitalization of both the warranty
  disclaimer and the liability limitation in Draft 3.

As noted by @fmarier, MPL 2.0 apparently found less annoying ways of
meeting the conspicuousness requirement. The irony (long noted by
Bradley Kuhn, as I recall) is that putting provisions in all caps
makes them less easy to read. I propose that we not worry about making
such provisions be "conspicuous" while GPL.next remains a draft
license, as this will simply make it harder to develop the content of
the provisions, but the questions of the extent to which the
conspicuousness requirement applies at all, and how it can be met,
should be explored.