Deletes penultimate paragraph of section 7.
I don't see what purpose this paragraph serves. If a licensee having
discretion to place additional terms on a work, or a portion of a
work, fails to do so, then there are no additional terms. If a
licensee having discretion to place additional terms on a work/portion
wishes to do so, the licensee will naturally be expected to provide
some customary notice of the applicability of such additional
terms. This is just unnecessary clutter.