Provides clearer definitions of "the Program" and "You"; "You"/"Your" now capitalized in accordance with typical legal practice.
Regarding the clarified definition of "the Program", this is motivated
by a concern raised not long after the final release of GPLv3 by,
separately, TI and Sun. The concern was that "the Program" could
somehow be read to mean "all possible GPLv3-licensed works in
existence" (which in turn raised fears about unbounded scope of
certain patent-related provisions). (The noted textual ambiguity has
counterparts in other free software licenses, and is probably
unavoidable on some level owing to limitations on the precision of
English.)

The FSF addressed this issue publicly in this document:
http://www.gnu.org/licenses/gplv3-the-program.html

The use of "particular" in my redefinition here reinforces this
reasonable interpretation.