Deleted "discriminatory patent licenses" provision.
The deleted paragraph is a relatively complex and narrowly drawn
provision which was introduced in GNU GPLv3 in response to the
Novell-Microsoft deal of late 2006. Oversimplifying a bit here, the
intent of the provision was to prospectively make it a violation of
the license to distribute pursuant to a deal with perceived similar
characteristics (for example, the granting of limited patent
nonasserts by a non-NPE third party to customers of the
distributor). The provision was necessarily drafted with only limited
knowledge of the details of the deal to which it was a response. The
provision was drafted with intentional narrowness out of concern that
it might reach varieties of vendor behavior unrelated to the issues
that motivated it.

This one calls for a return to the drawing board. One must first
soberly ascertain the proper scope of liberty-or-death to determine
whether this provision is even necessary or desirable even at a highly
general level. The narrowness of the provision makes its nonpolitical
value highly dubious. Moreover, if there is some flaw in
liberty-or-death, the clear solution is to amend liberty-or-death
directly.