Folds downstream-shielding-remnant provision into liberty-or-death.
The paragraph in question began its life in public GNU GPLv3 drafts as
a simple but broad "downstream shielding" requirement which had
implications for the scope of liberty-or-death. It was narrowed
substantially as a result of discussions with vendors during the GNU
GPLv3 drafting process.

The provision seems to state a kind of safe harbor in relation to
liberty-or-death. Therefore, it is sensible to include it as a
qualification to that section. If there is some further policy goal
worth exploring here it should be done through clarification of the
scope of liberty-or-death.