Made several relatively minor changes to section 5.
In section 0, "modified version" [of the Program] is defined
synonymously to "work based on the Program". The latter term was used
in GNU GPLv2 (though, curiously, it was defined more broadly). At
least in Copyleft.next at this point, the only place where "based on"
is used is at the beginning of section 5. I have replaced it with the
shorter "Modified Version", used conventional initial-capitalization
for "Modified Version", and, for clarity, have used "Modified Version"
where this section goes on to speak of the modified version as "the
work". Some other minor clarifications were made.

This commit also simplifies and clarifies the 5d "Appropriate Legal
Notices" clause.

Finally, the defined term "aggregate" in the last paragraph gets
conventional initial-capitalization.