Deletes ALN definition in sec. 0; liberalizes ALN requirement in sec. 5.
This commit introduces two changes. First, the definition of
"Appropriate Legal Notices" in section 0 is deleted. To my
recollection this factored-out definition originated with the
introduction of limited badgeware compatibility in late drafts of GNU
GPLv3. It is made unnecessary by the second change in this commit.

The predecessor GNU GPL versions of 5d were somewhat controversial in
some quarters because of their tension with the fundamental right of
modification. The rewording here attempts to recast it in way that is
intended to minimize the burdens on licensees who modify and
distribute. (There is a separate policy issue, which I do not consider
here, regarding whether any such requirement ought to be in the
license.) Particularly new here is the sentence calling for
construction in favor of the licensee's right to modify.

A subtler change here places further barriers to would-be badgeware
licensors. Replacing the ALN definition in section 0, a limited
category of "Legal Notices" is defined, which includes *textual*
author attributions. Thus requirements to preserve a graphical
"Powered by" logo (or worse) can no longer be grounded in the 5d
clause.