Various improvements.
Clarified clause stating distribution does not include pure network
services interaction.

Migrated 'contractor' clause of GNU GPLv3 section 2 (which had been
requested particularly by counsel for securities/financial services
industry interests during GNU GPLv3 drafting) to section 0 so that it
becomes a second explicit variety of non-distribution. A substantive
change was made here: the GNU GPLv3 version suggests that the
contractee-conveyor must comply fully with normal GNU GPLv3 conveying
requirements as to the portions of the work given to the
contractor/outsourced data center that are not copyrighted by the
conveyor.

This seems to me to be unlikely to be complied with in practice, and
perhaps relatively burdensome. Since, in a sense, the rationale for
the GNU GPLv3 provision in its section 2 was (roughly) that
contractors and outsourced data centers are reasonably treated like
employees, perhaps it should be easier for licensees to transfer works
to contractors and data centers without worrying about ordinary
distribution-triggered compliance. The policy issue needs to be
explored further, but for the time being I am proceeding with this
change.

The explicit prohibition on sublicensing is removed. As noted in GNU
GPLv3, the automatic licensing provision would seem to make
sublicensing superfluous. German lawyers Till Jaeger and Axel Metzger
argued during GNU GPLv3 drafting that GNU GPLv2 was undesirably
unclear in seeming to allow sublicensing by negative implication but
having the automatic licensing feature that would seem to be designed
to preclude sublicensing.

I recall being quite impressed with the force of this formal analysis
in 2006, but I now see it as one or two extra lines of text in the GNU
GPL that don't have to be there. GNU GPLv2 has hardly suffered in any
practical way as a result of this formal puzzle concerning
sublicensing. Even if sublicensing seems inconsistent with automatic
licensing, what is the real harm if some GNU GPL transaction is
structured as an act of sublicensing? An explicit prohibition on
sublicensing seems to unnecessarily intensify a tone of
restrictiveness.

The first two paragraphs of the Additional Terms section, discussing
additional permissions, are deleted in this commit in the interests of
advancing textual economy. Everything here seems to be an obvious
consequence of copyright law. GNU GPLv2 has not suffered through
absence of such an explicit framework; indeed the deleted material is
largely an effort to codify FSF practice and interpretation under GNU
GPLv2.