BitTorrent option of section 6 deleted.
GNU GPLv3 added a Corresponding Source compliance option focused
specifically on BitTorrent (though it was drafted generally as "peer
to peer transmission"), in response to a stet comment. This was much
noted as an important advance of GNU GPLv3 over GNU GPLv2.

But in reality, the value of this addition is rather
questionable. Compliance using BitTorrent or the like under GNU GPLv3
proceeds as it has done under GNU GPLv2; it is probably typically not
strictly in compliance with the clause set forth in GNU GPLv3. Thus,
viewed legalistically, we had noncompliance before and we have
noncompliance now, and in both cases no licensors bother to enforce
against this course of conduct. So, in the interest of advancing
textual economy, we should delete this clause.