More explicit statement of royalty-freeness of all granted licenses.
In GNU GPLv3, the only explicit statement that granted licenses are
royalty-free occurs by way of an example given to the "no further
restrictions" rule. This is unsatisfactory in appearing not to bind
the original licensor (cf. the way the patent license grant of GNU
GPLv3 is formulated).

GNU GPLv2 indicates that distributed works based on the Program must
be licensed "at no charge to all third parties", and the warranty
disclaimer is prepended with a statement that "the Program is licensed
free of charge". These features were removed from GNU GPLv3, I believe
primarily to remove any doubt about the right to charge any fee for
distribution of copies (see also the second paragraph of GNU GPLv3
section 4).

In this commit, we delete the "no further restrictions" example and
add language to the "basic permissions" section clarifying that all
license are granted royalty-free.