Removes "or no price" from 2nd para. of verbatim distribution section.
Since "any price" includes a price of zero, it is unnecessary to
explicitly state "no price". Moreover, the historical concern here
related to residual doubts about commercialization of free software;
thus if there is anything that needs to be said here it is that there
is no upper bound on charges for distribution.