Added definition of "Licensor".
Added explicit definition of "Licensor", adapted from CC 3.0, though
it isn't used yet. I antiticipate that such a definition may prove
useful in future changes. This commit also effectively reverts an
earlier change which had deleted a clarification that the licensee
("You") can be an individual or an organization ("entity" is used
now).