Various substantive & stylistic changes to definitions in section 0.
Restored a definition of "License" (with the intent of disambiguating
from uses of lower-case verb/noun "license").

One notable change is restoring the term "derivative work" (used, as
is well known, in GNU GPLv2, but abandoned in GNU GPLv3), but as a
defined term and with initial capitalization. The scope is intended to
be the same as "work based on the Program" in the GNU GPL and the way
in which I believe the drafters of GNU GPLv2 intended "derivative
work" to be understood. The definition is largely based on the
modify/based-on definition of GNU GPLv3 but greater clarity is
attempted; thus it replaces "Modified Version" of previous
Copyleft.next commits. (Therefore, uses of "Modified Version" in
section 5 have been replaced with "Derivative Work".)

The earliest discussion draft of GNU GPLv3 used "derivative work"
similarly to GNU GPLv2. The later avoidance of that term was a result
of substantial criticism of the GNU GPLv2 definition asserted by US
lawyers, criticism which had begun years before the GNU GPLv3 drafting
process. I believe I understand, today, the main thrust of that
criticism, but I have addressed it here in a way which I believe
fulfills the intent of RMS. In particular, the definition adds an
"avoidance of doubt" clause that makes clear that "Derivative Work"
could well be broader (or even narrower) in scope than "derivative
work" under US law. This was the intended effect of the reformulated
modify/based-on definitions in GNU GPLv3.

Another interesting change is that the "Aggregate" clause at the end
of section 5 (a more complex descendant of the simpler "mere
aggregation" clause of GNU GPLv2) is deleted, and instead a definition
of "Mere Aggregate" is defined and then included as an additional
carveout to the "Derivative Work" definition. (In the GNU GPL,
mere-aggregation/"Aggregate" was always, in a sense, a
for-the-avoidance-of-doubt limit on the definition of
derivative-work/work-based-on-the-Program/modified-version; making
"Mere Aggregate" a definition which then reduces the definitional
scope of "Derivative Work" makes this all clearer.) I have not carried
over what is sometimes called (by me) the "Larry Rosen clause" of the
GNU GPLv3 Aggregate paragraph (as it was a late change in response to
a late provocative interpretive statement by Mr. Rosen), as I do not
believe it adds any clarity to the definition. The feature of the GNU
GPLv3 Aggregate paragraph that is substantively different from the GNU
GPLv2 "mere aggregation" clause is what I believe is an effort to deal
with the possibility of denials of GNU GPL rights through restrictive
compilation licenses. The way this is worded in GNU GPLv3 can be
improved upon and I have tried to do that here in the "Mere Aggregate"
definition.

Another change is a positive definition of "Distribute", preceding the
existing negative definitions.